+++ tests/test_accuracy.py
@@ -37,16 +37,12 @@ _KNOWN_FAILURES: frozenset[str] = frozenset(
         "cp850-ms/culturax_00000.txt",
         "cp858-en/culturax_00000.txt",
         "cp858-ms/culturax_00000.txt",
-        "cp932-ja/y-moto.com.xml",
         "gb2312-zh/_mozilla_bug171813_text.html",
         "iso-8859-15-en/culturax_00002.txt",
         "iso-8859-16-hu/culturax_OSCAR-2019_82421.txt",
         "iso-8859-16-ro/_ude_1.txt",
         "macroman-en/culturax_mC4_84512.txt",
         "macroman-id/culturax_mC4_114889.txt",
-        "utf-8-en/finnish-utf-8-latin-1-confusion.html",
-        "utf-16be-zh/sample_chinese_no_bom.txt",
-        "utf-16le-zh/sample_chinese_no_bom.txt",
         "windows-1252-no/culturax_00002.txt",
     }
 )
@@ -59,14 +55,11 @@ _KNOWN_ERA_FILTERED_FAILURES: frozenset[str] = frozenset(
     {
         "cp500-es/culturax_mC4_87070.txt",
         "cp850-fi/culturax_00001.txt",
-        "cp932-ja/y-moto.com.xml",
         "gb2312-zh/_mozilla_bug171813_text.html",
         "iso-8859-2-hu/torokorszag.blogspot.com.xml",
         "iso-8859-16-hu/culturax_OSCAR-2019_82421.txt",
         "iso-8859-16-ro/_ude_1.txt",
         "macroman-da/culturax_mC4_83469.txt",
-        "utf-16be-zh/sample_chinese_no_bom.txt",
-        "utf-16le-zh/sample_chinese_no_bom.txt",
     }
 )